hw5/requirements.txt
numpy==1.12.1
scikit-learn==0.18.1
nltk==3.2.2
Cython==0.25.2
scipy==0.19.0
six==1.10.0
Theano==0.9.0
tensorflow-gpu==1.1.0
h5py==2.7.0
Keras==2.0.4

